Blog#new_post and Blog#postmaker=
Blog has something that can make posts which we can plug in. We specify
what sort of posts this thing creates. As this is Ruby we are specifying
that a post is any object with a property blog set to the blog.

Some interesting compact rspec. Also interesting block centric code from
Avdi.